# openapi.yaml
openapi: 3.1.0
info:
    title: cookd-api
    description: |
        A cooking coach. Send ingredients or photos, get recipes and advice streamed back from Claude.
        Sign in once (password, or Google if the server has it switched on), then call everything
        else with the session token as a bearer token.
    version: 1.0.0
servers:
    - url: https://cookd.dragunov.dev
security:
    - session: []
paths:
...
    /ingredients:
        get:
            operationId: listIngredients
            summary: Your pantry, alphabetically
            description: Everything you buy, in stock or not. Items stay when they run out.
            responses:
                "200":
                    description: The pantry
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/Ingredient" }
                "401": { $ref: "#/components/responses/Error" }
        post:
            operationId: addIngredients
            summary: Add ingredients from a pasted list
            description: |
                Splits `text` on commas, semicolons and line breaks. Names are lowercased and
                duplicates dropped. An ingredient already in the pantry is switched back **on**,
                so typing "milk" again means you bought milk. Answers with the whole pantry.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [text]
                            additionalProperties: false
                            properties:
                                text: { type: string, examples: ["eggs, rice\ntomatoes"] }
            responses:
                "200":
                    description: The pantry after the change
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/Ingredient" }
                "400": { $ref: "#/components/responses/Error" }
                "401": { $ref: "#/components/responses/Error" }
    /ingredients/{id}:
        parameters:
            - name: id
              in: path
              required: true
              schema: { type: integer, format: int64 }
        patch:
            operationId: updateIngredient
            summary: Switch an ingredient on or off
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [inStock]
                            additionalProperties: false
                            properties:
                                inStock: { type: boolean }
            responses:
                "200":
                    description: The ingredient
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Ingredient" }
                "400": { $ref: "#/components/responses/Error" }
                "401": { $ref: "#/components/responses/Error" }
                "404": { $ref: "#/components/responses/Error" }
        delete:
            operationId: deleteIngredient
            summary: Remove an ingredient from the pantry for good
            responses:
                "204": { description: Removed }
                "401": { $ref: "#/components/responses/Error" }
                "404": { $ref: "#/components/responses/Error" }
    /journal:
        get:
            operationId: listJournal
            summary: Dishes you have cooked, most recently cooked first
            description: Up to 200 entries.
            responses:
                "200":
                    description: The journal
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/JournalEntry" }
                "401": { $ref: "#/components/responses/Error" }
        post:
            operationId: addJournalEntry
            summary: Record a dish and how it went
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/NewJournalEntry" }
            responses:
                "201":
                    description: Saved
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/JournalEntry" }
                "400": { $ref: "#/components/responses/Error" }
                "401": { $ref: "#/components/responses/Error" }
    /journal/{id}:
        delete:
            operationId: deleteJournalEntry
            summary: Delete a journal entry
            parameters:
                - name: id
                  in: path
                  required: true
                  schema: { type: integer, format: int64 }
            responses:
                "204": { description: Deleted }
                "401": { $ref: "#/components/responses/Error" }
                "404": { $ref: "#/components/responses/Error" }
...
components:
    securitySchemes:
        session:
            type: http
            scheme: bearer
            description: The token from `POST /auth/login` or `POST /auth/google`
...
    responses:
        Error:
            description: Failure. The status code says which; the body carries a message.
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
    schemas:
        Ingredient:
            type: object
            required: [id, name, inStock]
            properties:
                id: { type: integer, format: int64 }
                name: { type: string, description: Lowercase }
                inStock: { type: boolean }
        JournalEntry:
            type: object
            required: [id, dish, cookedOn, taste, minutes, note, createdAt]
            properties:
                id: { type: integer, format: int64 }
                dish: { type: string }
                cookedOn: { type: string, format: date }
                taste: { type: integer, minimum: 1, maximum: 5 }
                minutes: { type: [integer, "null"], description: How long it took, if recorded }
                note: { type: string }
                createdAt: { type: string, format: date-time }
        NewJournalEntry:
            type: object
            required: [dish, taste]
            additionalProperties: false
            properties:
                dish: { type: string, maxLength: 100 }
                taste: { type: integer, minimum: 1, maximum: 5 }
                minutes: { type: integer, minimum: 1, maximum: 1440, description: How long it took, start to plate }
                note: { type: string, maxLength: 500 }
                cookedOn:
                    type: string
                    format: date
                    description: Defaults to today; may not be in the future
...
        Error:
            type: object
            required: [error]
            properties:
                error: { type: string }
...